+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+## Version 0.1.2 (2024-06-09)
+* Moved goup directory to a subdirectoy within GOPATH (juliaup style)
+* Added install script for easy updates
+* Switched to using anyhow internally
+
 ## Version 0.1.1 (2023-02-19)
 * Added progress bar during downloads
 * Changed some dependencies